Rename classes and fix variables
