+++ src/neo-build/Neo.Build.Core/Helpers/PrivateKeyHelper.cs
@@ -1,6 +1,6 @@
 // Copyright (C) 2015-2025 The Neo Project.
 //
-// KeyPairHelper.cs file belongs to the neo project and is free
+// PrivateKeyHelper.cs file belongs to the neo project and is free
 // software distributed under the MIT software license, see the
 // accompanying file LICENSE in the main directory of the
 // repository or http://www.opensource.org/licenses/mit-license.php
@@ -16,9 +16,9 @@ using System;
 using System.Security;
 using System.Security.Cryptography;
 
-namespace Neo.Build.Core.Cryptography
+namespace Neo.Build.Core.Helpers
 {
-    public static class KeyPairHelper
+    public static class PrivateKeyHelper
     {
         public static KeyPair CreateNew()
         {
